requirements.txt
Django==5.2.4
djangorestframework==3.15.2
django-filter==23.5
Pillow==11.3.0
python-dotenv==1.0.0
